Modify rename handle
