+++ api/memo_handler.go
@@ -8,6 +8,14 @@ import (
 
 	_ "github.com/lib/pq"
 
+	"bytes"
+	"fmt"
+	"log"
+	"strconv"
+	"strings"
+
+	"github.com/go-chi/chi"
+
 	"github.com/ddddddO/tag-mng/domain"
 )
 
@@ -87,5 +95,231 @@ func setColor(m *domain.Memo) {
 		m.RowVariant = "secondary"
 	case 5:
 		m.RowVariant = "success"
+	}
+}
+
+func MemoDetailHandler(DB *sql.DB) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		memoId := chi.URLParam(r, "id")
+		if len(memoId) == 0 {
+			errResponse(w, http.StatusBadRequest, "empty value 'memoId'", nil)
+			return
+		}
+
+		params := r.URL.Query()
+		userId := params.Get("userId")
+		if len(userId) == 0 {
+			errResponse(w, http.StatusBadRequest, "empty value 'userId'", nil)
+			return
+		}
+
+		const memoDetailQuery = `
+	SELECT
+	    m.id AS id,
+	    m.subject AS subject,
+		m.content AS content,
+		m.is_exposed AS is_exposed,
+		(SELECT jsonb_agg(t.id)
+		  FROM memos m
+		  JOIN memo_tag mt
+		  ON m.id = mt.memos_id
+		  JOIN tags t
+		  ON mt.tags_id = t.id
+	      WHERE m.id = $1 AND m.users_id = $2
+		) AS tag_ids,
+		(SELECT jsonb_agg(t.name)
+		  FROM memos m
+		  JOIN memo_tag mt
+		  ON m.id = mt.memos_id
+		  JOIN tags t
+		  ON mt.tags_id = t.id
+	      WHERE m.id = $1 AND m.users_id = $2
+		) AS tag_names
+		   FROM memos m
+		   JOIN memo_tag mt
+		   ON m.id = mt.memos_id
+		   JOIN tags t
+		   ON mt.tags_id = t.id
+		WHERE m.id = $1 AND m.users_id = $2
+		GROUP BY m.id
+	`
+
+		rows, err := DB.Query(memoDetailQuery, memoId, userId)
+		if err != nil {
+			errResponse(w, http.StatusInternalServerError, "failed to connect db 2", err)
+			return
+		}
+		rows.Next()
+		var (
+			memoDetail domain.Memo
+			tagIds     string
+			tagNames   string
+		)
+		// NOTE: 気持ち悪いけど、tagIds/tagNamesは別変数で取得して、sliceに変換してmemoDetailのフィールドに格納する
+		err = rows.Scan(
+			&memoDetail.ID, &memoDetail.Subject, &memoDetail.Content, &memoDetail.IsExposed,
+			&tagIds, &tagNames,
+		)
+		if err != nil {
+			errResponse(w, http.StatusInternalServerError, "failed to connect db 3", err)
+			return
+		}
+		memoDetail.TagIds = strToIntSlice(tagIds)
+		memoDetail.TagNames = strToStrSlice(tagNames)
+
+		//ref: https://qiita.com/shohei-ojs/items/311ef080cd5cff1e0e16
+		var memoDetailJson bytes.Buffer
+		encoder := json.NewEncoder(&memoDetailJson)
+		encoder.SetEscapeHTML(false)
+		encoder.Encode(memoDetail)
+
+		w.WriteHeader(http.StatusOK)
+		w.Write(memoDetailJson.Bytes())
+	}
+}
+
+func strToIntSlice(s string) []int {
+	if len(s) <= 2 {
+		return []int{}
+	}
+
+	ss := strings.Split(s[1:len(s)-1], ",")
+	var nums []int
+	for _, strNum := range ss {
+		strNum = strings.TrimSpace(strNum)
+		num, err := strconv.Atoi(strNum)
+		if err != nil {
+			panic(err)
+		}
+		nums = append(nums, num)
+	}
+	return nums
+}
+
+func strToStrSlice(s string) []string {
+	if len(s) <= 2 {
+		return []string{}
+	}
+
+	return strings.Split(s[1:len(s)-1], ",")
+}
+
+type UpdatedMemo struct {
+	UserId        int    `json:"user_id"`
+	MemoId        int    `json:"memo_id"`
+	MemoSubject   string `json:"memo_subject"`
+	MemoContent   string `json:"memo_content"`
+	MemoIsExposed bool   `json:"memo_is_exposed"`
+}
+
+func MemoUpdateHandler(DB *sql.DB) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		log.Print("----MemoUpdateHandler----")
+		var updatedMemo UpdatedMemo
+		if err := json.NewDecoder(r.Body).Decode(&updatedMemo); err != nil {
+			errResponse(w, http.StatusInternalServerError, "failed to unmarshal json", err)
+			return
+		}
+
+		const updateMemoQuery = `
+		UPDATE memos SET subject=$1, content=$2, is_exposed=$3
+		 WHERE id=$4 AND users_id=$5
+		`
+
+		result, err := DB.Exec(updateMemoQuery,
+			updatedMemo.MemoSubject, updatedMemo.MemoContent, updatedMemo.MemoIsExposed, updatedMemo.MemoId, updatedMemo.UserId,
+		)
+		if err != nil {
+			errResponse(w, http.StatusInternalServerError, "failed to connect db 2", err)
+			return
+		}
+		n, err := result.RowsAffected()
+		if err != nil {
+			errResponse(w, http.StatusInternalServerError, "failed to connect db 3", err)
+			return
+		}
+		if n != 1 {
+			errResponse(w, http.StatusInternalServerError, "failed to update memo", nil)
+			return
+		}
+		w.WriteHeader(http.StatusOK)
+	}
+}
+
+type CreatedMemo struct {
+	UserId      int    `json:"user_id"`
+	TagIds      []int  `json:"tag_ids"`
+	MemoSubject string `json:"memo_subject"`
+	MemoContent string `json:"memo_content"`
+}
+
+func MemoCreateHandler(DB *sql.DB) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		log.Print("----MemoCreateHandler----")
+		var createdMemo CreatedMemo
+		if err := json.NewDecoder(r.Body).Decode(&createdMemo); err != nil {
+			errResponse(w, http.StatusInternalServerError, "failed", err)
+			return
+		}
+
+		var createMemoQuery = `
+WITH inserted AS (INSERT INTO memos(subject, content, users_id) VALUES($1, $2, $3) RETURNING id)
+INSERT INTO memo_tag(memos_id, tags_id) VALUES
+	((SELECT id FROM inserted), 1)
+	%s;
+`
+
+		var valuesStr string
+		for _, tagId := range createdMemo.TagIds {
+			valuesStr += fmt.Sprintf(",((SELECT id FROM inserted), %d)", tagId)
+		}
+		createMemoQuery = fmt.Sprintf(createMemoQuery, valuesStr)
+
+		_, err := DB.Exec(createMemoQuery,
+			createdMemo.MemoSubject, createdMemo.MemoContent, createdMemo.UserId,
+		)
+		if err != nil {
+			errResponse(w, http.StatusInternalServerError, "failed to connect db 2", err)
+			return
+		}
+		w.WriteHeader(http.StatusOK)
+	}
+}
+
+type DeleteMemo struct {
+	UserId int `json:"user_id"`
+	MemoId int `json:"memo_id"`
+}
+
+func MemoDeleteHandler(DB *sql.DB) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		log.Print("----MemoDeleteHandler----")
+		var deleteMemo DeleteMemo
+		if err := json.NewDecoder(r.Body).Decode(&deleteMemo); err != nil {
+			errResponse(w, http.StatusInternalServerError, "failed", err)
+			return
+		}
+
+		const deleteMemoQuery = `
+DELETE FROM memos WHERE users_id = $1 AND id = $2;
+`
+
+		result, err := DB.Exec(deleteMemoQuery,
+			deleteMemo.UserId, deleteMemo.MemoId,
+		)
+		if err != nil {
+			errResponse(w, http.StatusInternalServerError, "failed to connect db 2", err)
+			return
+		}
+		n, err := result.RowsAffected()
+		if err != nil {
+			errResponse(w, http.StatusInternalServerError, "failed to connect db 3", err)
+			return
+		}
+		if n != 1 {
+			errResponse(w, http.StatusInternalServerError, "failed", nil)
+			return
+		}
+		w.WriteHeader(http.StatusOK)
 	}
 }
